@@ -1,8 +1,8 @@
 package com.mindorks.demo;
 
 import android.os.Bundle;
-import android.support.annotation.Nullable;
-import android.support.v7.app.AppCompatActivity;
+import androidx.annotation.Nullable;
+import androidx.appcompat.app.AppCompatActivity;
 import android.util.Log;
 
 import com.mindorks.butterknifelite.ButterKnifeLite;
