@@ -3,7 +3,7 @@ package com.mindorks.demo;
 import android.os.Bundle;
 import android.os.Handler;
 import android.os.Looper;
-import android.support.v7.app.AppCompatActivity;
+import androidx.appcompat.app.AppCompatActivity;
 import android.util.Log;
 
 import com.mindorks.butterknifelite.ButterKnifeLite;
