@@ -1,7 +1,8 @@
 package com.mindorks.demo.gallery;
 
 import android.content.Context;
-import android.support.v7.widget.LinearLayoutManager;
+
+import androidx.recyclerview.widget.LinearLayoutManager;
 
 import com.mindorks.demo.Image;
 import com.mindorks.demo.R;
