@@ -2,7 +2,7 @@ package com.mindorks.placeholderview.widget;
 
 
 import android.content.Context;
-import android.support.annotation.NonNull;
+import androidx.annotation.NonNull;
 import android.util.AttributeSet;
 import android.widget.EditText;
 
