@@ -1,7 +1,5 @@
 package com.mindorks.placeholderview.annotations;
 
-import android.support.annotation.IdRes;
-
 import java.lang.annotation.ElementType;
 import java.lang.annotation.Retention;
 import java.lang.annotation.RetentionPolicy;
@@ -13,5 +11,5 @@ import java.lang.annotation.Target;
 @Retention(RetentionPolicy.SOURCE)
 @Target(ElementType.METHOD)
 public @interface LongClick {
-    @IdRes int value();
+    int value();
 }
